test: consolidate AutoNovelAgent tests
