+++ tests/test_auto_novel_agent.py
@@ -11,6 +11,13 @@ def reset_engines():
     AutoNovelAgent.SUPPORTED_ENGINES = {"unity", "unreal"}
 
 
+def test_supports_engine_case_insensitive():
+    agent = AutoNovelAgent()
+    assert agent.supports_engine("UNITY")
+    assert agent.supports_engine("unreal")
+    assert not agent.supports_engine("godot")
+
+
 def test_add_supported_engine_enables_creation():
     agent = AutoNovelAgent()
     with pytest.raises(ValueError):
@@ -25,3 +32,30 @@ def test_remove_supported_engine_blocks_creation():
     agent.remove_supported_engine("unity")
     with pytest.raises(ValueError):
         agent.create_game("unity")
+
+
+def test_create_game_disallows_weapons():
+    agent = AutoNovelAgent()
+    with pytest.raises(ValueError, match="Weapons are not allowed"):
+        agent.create_game("unity", include_weapons=True)
+
+
+def test_create_game_rejects_unsupported_engine():
+    agent = AutoNovelAgent()
+    with pytest.raises(ValueError):
+        agent.create_game("cryengine")
+
+
+def test_generate_story_requires_theme():
+    agent = AutoNovelAgent()
+    with pytest.raises(ValueError):
+        agent.generate_story("")
+
+
+def test_generate_story_series_produces_multiple_stories():
+    agent = AutoNovelAgent()
+    stories = agent.generate_story_series(["mystery", "space"], protagonist="Explorer")
+    assert len(stories) == 2
+    assert "Explorer" in stories[0]
+    assert "mystery" in stories[0]
+    assert "space" in stories[1]
